DFE-968 - refactored service structure
