+++ src/explore-education-statistics-admin/src/services/common/types/types.ts
@@ -1,5 +1,5 @@
-import { ReleaseApprovalStatus } from '@admin/services/api/dashboard/types';
-import { User } from '../../../PrototypeLoginService';
+import { ReleaseApprovalStatus } from '@admin/services/dashboard/types';
+import { User } from '@admin/services/PrototypeLoginService';
 
 export interface IdLabelPair {
   id: string;
