+++ src/explore-education-statistics-admin/src/services/dashboard/service.ts
@@ -1,8 +1,8 @@
-import mocks from '@admin/services/api/edit-release/data/mock/axios-mock';
-import { DataFileView } from '@admin/services/api/edit-release/data/types';
 import Client from '@common/services/api/Client';
 import { commaSeparated } from '@common/services/util/paramSerializers';
 import axios from 'axios';
+import mocks from './mock/axios-mock';
+import { AdminDashboardPublication, ThemeAndTopics } from './types';
 
 const createClient = async () => {
   const baseUrl = process.env.CONTENT_API_BASE_URL;
@@ -10,12 +10,6 @@ const createClient = async () => {
   const axiosInstance = axios.create({
     baseURL: `${baseUrl}/api/`,
     paramsSerializer: commaSeparated,
-  });
-
-  axios.interceptors.request.use(request => {
-    // eslint-disable-next-line no-console
-    console.log('Starting Request', request);
-    return request;
   });
 
   const decoratedAxios =
@@ -29,9 +23,19 @@ const createClient = async () => {
 const apiClient = createClient();
 
 export default {
-  getReleaseDataFiles(releaseId: string): Promise<DataFileView> {
+  getThemesAndTopics(userId: string): Promise<ThemeAndTopics[]> {
     return apiClient.then(client =>
-      client.get<DataFileView>(`/release/${releaseId}/datafiles`),
+      client.get<ThemeAndTopics[]>('/Themes', { params: { userId } }),
+    );
+  },
+  getPublicationsByTopic(
+    topicId: string,
+    userId: string,
+  ): Promise<AdminDashboardPublication[]> {
+    return apiClient.then(client =>
+      client.get<AdminDashboardPublication[]>('/Publications', {
+        params: { topicId, userId },
+      }),
     );
   },
 };
